ajax only sub follow request

+++ project/content/views.py
@@ -98,31 +98,37 @@ class SubPostListView(ListView):
 
 
 class SubFollowCreate(View):
-	def get(self, *args, **kwargs):
+	def get(self, request, *args, **kwargs):
 		sub_followed = self.kwargs['followed']
 
-		sub_followed_obj = SubFollow.objects.create(follower=self.request.user,sub_id=sub_followed)
-		sub_followed_obj.save()
-		sub_followed_obj.follower.sub_following_number += 1
-		sub_followed_obj.follower.save()
-		sub_followed_obj.sub.follower_number += 1
-		sub_followed_obj.sub.save()
+		if request.is_ajax():
+			sub_followed_obj = SubFollow.objects.create(follower=self.request.user,sub_id=sub_followed)
+			sub_followed_obj.save()
+			sub_followed_obj.follower.sub_following_number += 1
+			sub_followed_obj.follower.save()
+			sub_followed_obj.sub.follower_number += 1
+			sub_followed_obj.sub.save()
 
-		return HttpResponse(status=200)
+			return HttpResponse(status=200)
+		else:
+			return HttpResponseRedirect('/sub/%s' % sub_unfollowed)
 
 
 class SubFollowDelete(View):
-	def get(self, *args, **kwargs):
+	def get(self, request, *args, **kwargs):
 		sub_unfollowed = self.kwargs['unfollowed']
 
-		sub_unfollowed_obj = SubFollow.objects.get(follower=self.request.user, sub_id=sub_unfollowed)
-		sub_unfollowed_obj.follower.sub_following_number -= 1
-		sub_unfollowed_obj.follower.save()
-		sub_unfollowed_obj.sub.follower_number -= 1
-		sub_unfollowed_obj.sub.save()
-		sub_unfollowed_obj.delete()
+		if request.is_ajax():
+			sub_unfollowed_obj = SubFollow.objects.get(follower=self.request.user, sub_id=sub_unfollowed)
+			sub_unfollowed_obj.follower.sub_following_number -= 1
+			sub_unfollowed_obj.follower.save()
+			sub_unfollowed_obj.sub.follower_number -= 1
+			sub_unfollowed_obj.sub.save()
+			sub_unfollowed_obj.delete()
 
-		return HttpResponse(status=200)
+			return HttpResponse(status=200)
+		else:
+			return HttpResponseRedirect('/sub/%s' % sub_unfollowed)
 
 
 class PostCommitView(CreateView):
